requirements.txt
tensorflow
tensorflow-probability
scipy
flask
flask-session
forms
numpy
pandas
flask-bootstrap
hashlib